remove rx and other old stuff
