+++ ember-cli-build.js
@@ -41,7 +41,5 @@ module.exports = function(defaults) {
     behave. You most likely want to be modifying `./index.js` or app's build file
   */
 
-  app.import('vendor/dummy-deps/rx.js');
-
   return app.toTree();
 };
