--- tests/dummy/app/docs/examples/observables/template.hbs
@@ -1,32 +1,0 @@
-<h3>Experimental: Observables</h3>
-
-<p>
-  <strong>ember-concurrency</strong> integrates with Observables
-  and lets you apply the same concurrency constraints to
-  subscriptions to Observables that you could apply to a task &mdash;
-  just as you can configure a Task to <code>drop()</code> new
-  <code>.perform()</code>s while the task is already running, so can
-  you configure the subscription to an async sequence of events
-  (the Observable) to <code>drop()</code> events while the
-  previous event is being handled (or use any other task modifier
-  you like).
-</p>
-
-<h3>Live Example</h3>
-
-<h4>Nested loop of Observable range of numbers</h4>
-
-<button onclick={{perform computeStuff}}>Compute Stuff</button>
-
-<pre>{{#each values as |v|}}{{v.message}}{{/each}}</pre>
-
-{{code-snippet name="observables-timetable.js"}}
-
-<h4>Subscribing to Observable of Evented events</h4>
-
-<h5>Status: {{fooStatus}}</h5>
-
-<button {{action 'triggerFoo'}}>Trigger Foo</button>
-
-{{code-snippet name="observables-evented.js"}}
-
