+++ tests/unit/external/buffering-test.js
@@ -1,30 +1,9 @@
+/*
 import EmberObject from '@ember/object';
 import { later, run, join } from '@ember/runloop';
 import { task, timeout } from 'ember-concurrency';
 import { module, test } from 'qunit';
 
-const Observable = window.Rx.Observable;
-
-const {
-  range,
-  just,
-} = Observable;
-
-// 1 2 3 4 5 (150ms) 101 102 103 104 105
-let rangeObservable = Observable.concat(
-  range(1,5),
-  range(101,5).delay(150)
-);
-
-// 1 (2ms) 2 3 (100ms) 4 5 6
-let sporadicObservable = Observable.concat(
-  just(1),
-  just(2).delay(5),
-  just(3),
-  just(4).delay(100),
-  just(5),
-  just(6)
-);
 
 module('Unit: buffering', function() {
   function doBufferingTest(description, observable, bufferPolicyFn, expectations, maxConcurrency) {
@@ -105,3 +84,4 @@ module('Unit: buffering', function() {
   doBufferingTest("restartable(3): ranges",   rangeObservable,     (t) => t.restartable().maxConcurrency(3), [3,4,5,103,104,105], 5);
   doBufferingTest("restartable(3): sporadic", sporadicObservable,  (t) => t.restartable().maxConcurrency(3), [1,2,3,4,5,6], 3);
 });
+*/
