+++ tests/unit/external/scheduler-policies/helpers.js
@@ -1,0 +1,4 @@
+export function testScheduler(policy, args) {
+  let reducer = policy.makeReducer(...args);
+  return [...Array(10)].map(() => reducer.step());
+}
